Organization and bug fixes
Outside of any brackets, the text will be printed as is including empty spaces. But inside brackets, empty lines that come before or after brackets will be removed. This is necessary to make it fit with the python style of organization. 

Now an extra new line is not created at the beginning of the new file. 

There is also now only 1 function which comprises a bulk of the code and accounts for indentations, lack of indentations, and the first line issue.